requirements.txt
pydub
pyyaml
runez
sox
yt-dlp
